perf: optimize development server performance
- Extract Material-UI theme provider to reduce duplicate CssVarsProvider instances
- Optimize font loading for development environment to skip heavy font files
- Add Vite optimization for Material-UI components and icons
- Consolidate theme configuration in dedicated ThemeProvider component
- Configure code splitting for better bundle management
- Rename astro.config.mjs to astro.config.mts for TypeScript support
