+++ astro.config.mts
@@ -1,8 +1,7 @@
 import { defineConfig } from 'astro/config';
 import react from "@astrojs/react";
-
 import vercel from "@astrojs/vercel/static";
-import { setupMicroCMS } from './src/utils/integration';
+import { setupMicroCMS } from "./src/utils/integration";
 
 // https://astro.build/config
 export default defineConfig({
@@ -17,6 +16,38 @@ export default defineConfig({
       enabled: true,
     },
     imageService: true,
-    devImageService: true
+    devImageService: 'squoosh',
   }),
+  vite: {
+    optimizeDeps: {
+      include: [
+        '@mui/joy',
+        '@mui/icons-material/MoreVert',
+        '@mui/icons-material/Facebook',
+        '@mui/icons-material/Instagram',
+        '@mui/icons-material/GitHub',
+        '@mui/icons-material/Twitter',
+      ],
+      exclude: [
+        '@fontsource/noto-serif-jp',
+      ],
+    },
+    build: {
+      rollupOptions: {
+        output: {
+          manualChunks: {
+            'mui-core': ['@mui/joy'],
+            'mui-icons': [
+              '@mui/icons-material/MoreVert',
+              '@mui/icons-material/Facebook',
+              '@mui/icons-material/Instagram',
+              '@mui/icons-material/GitHub',
+              '@mui/icons-material/Twitter',
+            ],
+            'vendor': ['react', 'react-dom'],
+          },
+        },
+      },
+    },
+  },
 });
